Move <ChatInput> out of <Chat> and into its own file in Inputs
