+++ src/Components/Chats/Chat.js
@@ -16,20 +16,6 @@
     <Chat/>'s Parents:
       - <ChatPopup/>
 
-
-  2) Chat also contains <ChatInput/>, which is responsible for:
-    - being the text input that the user types into when using <Chat/>
-    - combining a <textarea/> with emoji selection tools
-
-    <ChatInput/>'s Props:
-      - n/a for now
-
-    <ChatInput/>'s Children:
-      - <AdjustableTextArea/>
-      - <EmojiSelect/>
-
-    <ChatInput/>'s Parents:
-      - <Chat/>
 */
 
 // Imports ---------------------------------------------------------------------
@@ -37,12 +23,8 @@
 import React from 'react';
 import './Chat.css';
 
-// npm
-import RenderEmoji from 'react-easy-emoji';
-
 // components
-import AdjustableTextArea from '../Inputs/AdjustableTextArea.js';
-import EmojiSelect        from '../Inputs/EmojiSelect.js';
+import ChatInput from '../Inputs/ChatInput.js';
 
 // =============================================================================
 // <Chat/>
@@ -61,86 +43,5 @@ class Chat extends React.Component {
     );
   }
 }
-
-
-// =============================================================================
-// <ChatInput/>
-// =============================================================================
-
-class ChatInput extends React.Component {
-
-  constructor() {
-    super();
-    this.state = {
-      text              : "",
-      emojiSelectActive : false
-    };
-    this.childTextArea = React.createRef(); // <- create this ref so we can call <AdjustableTextArea/>.addSubstringToText() for emojis
-  }
-
-  // sends the user's message and wipes state
-  sendText = (textToSend) => {
-    console.log(`sending: ${textToSend}`);
-    this.setState({text: "", emojiSelectActive: false});
-  }
-
-  onClick_toggleEmojiSelect = () => {
-    if (this.state.emojiSelectActive === true) {
-      this.childTextArea.current.focusOnTextArea();
-    }
-    this.setState({emojiSelectActive: !this.state.emojiSelectActive});
-  }
-
-  addEmojiToText = (emojiUnicode) => {
-    let newText = this.childTextArea.current.addSubstringToText(emojiUnicode);
-    this.setState({text: newText, emojiSelectActive: false});
-    this.childTextArea.current.focusOnTextArea();
-  }
-
-
-
-  // render --------------------------------------------------------------------
-
-  renderEmojiSelect = () => {
-
-    if (this.state.emojiSelectActive === false) {
-      return (
-        <div id="emoji-select-anchor"></div>
-      )
-    }
-
-    return (
-      <div id="emoji-select-anchor">
-        <div id="emoji-select-backdrop" onMouseLeave={(e) => e.stopPropagation()} onClick={this.onClick_toggleEmojiSelect}></div>
-        <div id="emoji-select-container">
-          <EmojiSelect selectEmoji={this.addEmojiToText}/>
-        </div>
-      </div>
-    );
-  }
-
-
-  render() {
-    let renderEmojiButtonText = (this.state.emojiSelectActive) ? '😀' : '🙂';
-    return (
-      <div id="ChatInput">
-        <div id="text-input-row">
-          <div id="text-area-container">
-            <AdjustableTextArea
-              ref        = {this.childTextArea}
-              text       = {this.state.text}
-              updateText = {(newText) => {this.setState({text: newText})}}
-              submitText = {this.sendText}
-              />
-          </div>
-          {(this.state.text.length === 0 && false) && <div id="attachments-button"></div>}
-          <div id="emojis-button" onClick={this.onClick_toggleEmojiSelect}>{RenderEmoji(renderEmojiButtonText)}</div>
-          {this.renderEmojiSelect()}
-        </div>
-      </div>
-    );
-  }
-}
-
 
 export default Chat;
